Alex Bakke - Portfolio Webpage
This is a public portfolio page that encompasses a small amount of the work I've done. A good majority is unavailable due to non disclosure agreements.

It does include a recent update to show application of a public API on Public Art in Asheville.

It also does not show off a major body of learning I have spent a lot of time doing: T-SQL / MySQL.

Thank you for the opportunity to apply!